+++ DTOs/ProdutoResponseDto.cs
@@ -5,7 +5,7 @@ public sealed class ProdutoResponseDto
     public int Id { get; set; }
     public string Nome { get; set; } = string.Empty;
     public string Descricao { get; set; } = string.Empty;
-    public double PrecoBase { get; set; }
+    public decimal PrecoBase { get; set; }
     public string? Categoria { get; set; }
     public string Status { get; set; } = string.Empty;
 }
